Revert "Finish adding translate and map API to website"
