+++ portfolio/pom.xml
@@ -35,12 +35,6 @@
       <artifactId>appengine-api-1.0-sdk</artifactId>
       <version>1.9.59</version>
     </dependency>
-
-    <dependency>
-      <groupId>com.google.cloud</groupId>
-      <artifactId>google-cloud-translate</artifactId>
-      <version>1.70.0</version>
-    </dependency>
   </dependencies>
 
   <build>
@@ -52,7 +46,7 @@
         <artifactId>appengine-maven-plugin</artifactId>
         <version>2.2.0</version>
         <configuration>
-          <deploy.projectId>sasymer-step</deploy.projectId>
+          <deploy.projectId>sasymer-step-2020</deploy.projectId>
           <deploy.version>1</deploy.version>
         </configuration>
       </plugin>
