--- portfolio/src/main/java/com/google/sps/servlets/TranslateServlet.java
@@ -1,35 +1,0 @@
-
-package com.google.sps.servlets;
-
-import com.google.cloud.translate.Translate;
-import com.google.cloud.translate.TranslateOptions;
-import com.google.cloud.translate.Translation;
-import java.io.IOException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/translate")
-public class TranslateServlet extends HttpServlet {
-
-  @Override
-  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    // Get the request parameters.
-    String originalText = request.getParameter("text");
-    System.out.println(originalText);
-    String languageCode = "en";
-
-    // Do the translation.
-    Translate translate = TranslateOptions.getDefaultInstance().getService();
-    Translation translation =
-        translate.translate(originalText, Translate.TranslateOption.targetLanguage(languageCode));
-    String translatedText = translation.getTranslatedText();
-    System.out.println(translatedText);
-
-    // Output the translation.
-    response.setContentType("text/html; charset=UTF-8");
-    response.setCharacterEncoding("UTF-8");
-    response.getWriter().println(translatedText);
-  }
-}
