+++ portfolio/src/main/webapp/cell_sim.html
@@ -3,20 +3,20 @@
   <head>
     <meta charset="UTF-8">
     <title>Cell Simulation</title>
-    <link rel="stylesheet" href="../style.css">
+    <link rel="stylesheet" href="style.css">
     <script src="script.js"></script>
   </head>
   <body>
     <div class="header">
-      <header class="paddedHeader">
-        <div style="float:left">
+      <header style="padding-bottom: 50px; padding-top: 50px">
+        <div style="float: left">
           <a href="index.html" class="navButton">Sanna Symer</a>
         </div>
-        <div style="float:right">
-          <a href="../aboutme.html" class="navButton">Me</a>
-          <a href="../fun.html" class="navButton">Fun</a>
-          <a href="../projects.html" class="navButton">Projects</a>
-          <a href="../comment-page/comments.html" class="navButton">Comments</a>
+        <div style="float: right">
+          <a href="aboutme.html" class="navButton">Me</a>
+          <a href="fun.html" class="navButton">Fun</a>
+          <a href="projects.html" class="navButton">Projects</a>
+          <a href="comments.html" class="navButton">Comments</a>
         </div>
       </header>
     </div>
