--- portfolio/src/main/webapp/chinese-page/translateEssay.js
@@ -1,19 +1,0 @@
-function doTranslate() {
-  console.log('translating');
-  const text = document.getElementById('chinese-paragraph').textContent;
-  console.log(text);
-
-  const resultContainer = document.getElementById('english-paragraph');
-  resultContainer.innerText = 'Loading...';
-
-  const params = new URLSearchParams();
-  params.append('text', text);
-
-  fetch('/translate', {
-    method: 'POST',
-    body: params
-  }).then(response => response.text())
-  .then((translatedMessage) => {
-    resultContainer.innerText = translatedMessage;
-  });
-}
