--- portfolio/src/main/webapp/chinese-page/translateStyle.css
@@ -1,10 +1,0 @@
-.paragraph-box {
-  text-align:justify; 
-  padding-left: 10%; 
-  padding-right: 10%;
-}
-
-.button-holder {
-  padding-left: 10%;
-  padding-right: 10%;
-}
