+++ portfolio/src/main/webapp/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="UTF-8">
-    <title>Sanna Symer</title>
+    <title>My Portfolio</title>
     <link rel="stylesheet" href="style.css">
     <script src="script.js"></script>
   </head>
@@ -27,7 +27,7 @@
         <p class="frontpage-paragraph">
             Hi, I'm <span class="frontpage-name">Sanna:</span> 
             a creative problem solver, loyal friend, and 
-            proud vegetarian.
+            proud vegetarian. 
             I'm interested in data science, machine learning, and 
             software engineering--especially in their applications 
             in the social sciences.
