+++ portfolio/src/main/webapp/projects.html
@@ -38,22 +38,22 @@
       </div>
 
       <div name="icons"> 
-        <a href="/project-pages/cell_sim.html">
+        <a href="cell_sim.html">
           <img src="/icons/slider.png" width=150px; height=150px; onmouseover="this.src='/icons/sim_hover.png';" 
           onmouseout="this.src='icons/slider.png';" alt="Sliders" style="padding:20px;">
         </a>
 
-        <a href="/project-pages/slogo.html">
+        <a href="slogo.html">
           <img src="/icons/turtlefd.png" width=150px; height=150px; onmouseover="this.src='/icons/turtle_hover.png';" 
           onmouseout="this.src='icons/turtlefd.png';" alt="Turtle" style="padding:20px;">
         </a>
 
-        <a href="/project-pages/card_engine.html">
+        <a href="card_engine.html">
           <img src="/icons/card.png" width=150px; height=150px; onmouseover="this.src='/icons/cards_hover.png';" 
           onmouseout="this.src='icons/card.png';" alt="Cards" style="padding:20px;">
         </a>
 
-        <a href="/project-pages/monopoly.html">
+        <a href="monopoly.html">
           <img src="/icons/monopoly.png" width=150px; height=150px; onmouseover="this.src='/icons/monopoly_hover.png';" 
           onmouseout="this.src='icons/monopoly.png';" alt="Monopoly" style="padding:20px;">
         </a>
@@ -95,8 +95,7 @@
           Click on some of the links below to see some "hidden" pages of this site.
         </p>
         <a href="experiment.html" class="personalLink">Bootstrap Experiments</a>
-        <a href="/map-page/map.html" class="personalLink">Map API</a>
-        <a href="/chinese-page/chinese.html" class="personalLink">Translate API</a>
+        <a href="map.html" class="personalLink">Map API</a>
       </div> 
     </div>
