+++ walkthroughs/week-4-libraries/blobstore/examples/hello-world/src/main/java/com/google/sps/servlets/HomeServlet.java
@@ -48,8 +48,6 @@ public class HomeServlet extends HttpServlet {
     // This is not how you'd do this in a real codebase!
     // See the hello-world-jsp or hello-world-fetch examples for more info.
     PrintWriter out = response.getWriter();
-    out.println("<p>" + uploadUrl + "</p>");
-
     out.println(
         "<form method=\"POST\" enctype=\"multipart/form-data\" action=\"" + uploadUrl + "\">");
