+++ walkthroughs/week-4-libraries/maps/examples/hello-world/src/main/webapp/index.html
@@ -2,7 +2,7 @@
 <html>
 <head>
   <title>Hello Google Maps</title>
-  <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyCImVHTwoNt0c7rw3DtNhyxLtKmo91ZiEw"></script>
+  <script src="https://maps.googleapis.com/maps/api/js?key=YOUR_API_KEY"></script>
   <script src="script.js"></script>
   <link rel="stylesheet" href="style.css">
 </head>
