+++ walkthroughs/week-5-tdd/intro/src/test/java/com/google/sps/GreeterTest.java
@@ -30,24 +30,4 @@ public final class GreeterTest {
 
     Assert.assertEquals("Hello Ada", greeting);
   }
-
-  @Test
-  public void testGreetingTrimsWhitespace() {
-    Greeter greeter = new Greeter();
-
-    String greeting = greeter.greet("   Ada   ");
-
-    // Whitespace should be trimmed
-    Assert.assertEquals("Hello Ada", greeting);
-  }
-
-  @Test
-  public void testGreetingTrimsSymbols() {
-    Greeter greeter = new Greeter();
-
-    String greeting = greeter.greet("Ada#5!");
-
-    //Non-letter chars should be deleted
-    Assert.assertEquals("Hello Ada", greeting);
-  }
 }
